Add first-pass script to show home games
I can get hold of a feed from the club's website, but that shows all
games: I only care about the ones that happen in Cambridge.

The script currently reads from `stdin` and writes to `stdout`: for the
next bit, hopefully I'll be able to grab the fixture list from the
Cambridge United website and spit it out over HTTP, so it'll integrate
with Google Calendar.